filtering : big refactor : put each filter in its own file
